renamed mapping to sensing. got reference detections going
