--- src/perception/sim_mapper/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/sim_mapper
-[install]
-install_scripts=$base/lib/sim_mapper
