+++ src/perception/sim_sensing/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>sim_mapper</name>
+  <name>sim_sensing</name>
   <version>0.0.0</version>
   <description>Outputs true cone locations in car frame of reference from sim map</description>
   <maintainer email="team@qutmotorsport.com">Alastair Bradford</maintainer>
