+++ src/perception/sim_sensing/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/sim_sensing
+[install]
+install_scripts=$base/lib/sim_sensing
